# optimize block_ref creation