The GORM create helpers dont have any way to avoid 'returning' the ID values of created records, and that was causing a noticeable performance hit in this critical section.