@whyrusleeping 2023-06-26
Hard to notice too much of a difference in the existing benchmarks because we're using an in-memory sqlite instance, but heres the numbers:
Before:
```
why@sirius ~/c/indigo (main)> go test ./carstore -bench=.
7256
goos: linux
goarch: amd64
pkg: github.com/bluesky-social/indigo/carstore
cpu: AMD Ryzen Threadripper 3970X 32-Core Processor
BenchmarkRepoWritesCarstore-64    	    2698	    535345 ns/op
BenchmarkRepoWritesFlatfs-64      	    2444	    540607 ns/op
BenchmarkRepoWritesSqlite-64      	    5752	    339228 ns/op
PASS
ok  	github.com/bluesky-social/indigo/carstore	4.878s
```

After:
```
why@sirius ~/c/indigo (feat/carstore-opti)> go test ./carstore -bench=.
9931
goos: linux
goarch: amd64
pkg: github.com/bluesky-social/indigo/carstore
cpu: AMD Ryzen Threadripper 3970X 32-Core Processor
BenchmarkRepoWritesCarstore-64    	    3488	    483285 ns/op
BenchmarkRepoWritesFlatfs-64      	    2767	    509982 ns/op
BenchmarkRepoWritesSqlite-64      	    6513	    326489 ns/op
PASS
ok  	github.com/bluesky-social/indigo/carstore	5.370s
```